+++ src/client/browser-detection.js
@@ -1,0 +1,7 @@
+'use strict';
+
+var isIe = require('@braintree/browser-detection/is-ie');
+
+module.exports = {
+  isIe: isIe
+};